Refactor and ping + status endpoints to separate impl
