+++ src/main/openapi/ds-discover-openapi_v1.yaml
@@ -25,7 +25,7 @@ paths:
   /monitor/ping:
     get:
       tags:
-        - '${project.name}'
+        - Service
       summary: 'Ping the server to check if the server is reachable.'
       operationId: ping
       x-useGenericResponse: false # If enabled the method will return Response
@@ -33,18 +33,6 @@ paths:
       responses:
         '200':
           description: 'OK'
-          content:
-            text/plain:
-              schema:
-                type: string
-        '406':
-          description: 'Not Acceptable'
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Error'
-        '500':
-          description: 'Internal Error'
           content:
             text/plain:
               schema:
@@ -62,7 +50,7 @@ paths:
   /monitor/status:
     get:
       tags:
-        - '${project.name}'
+        - Service
       summary: 'Detailed status / health check for the service'
       operationId: status
       responses:
